Reduce code duplication in received message handling
Move more of the old tracing also intro the received call targets

Factor our redundant code
