+++ src/som/interpreter/actors/ReceivedMessage.java
@@ -7,11 +7,9 @@ import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.DirectCallNode;
 
-import som.interpreter.SArguments;
 import som.interpreter.SomException;
 import som.interpreter.SomLanguage;
 import som.interpreter.nodes.MessageSendNode.AbstractMessageSendNode;
-import som.vm.VmSettings;
 import som.vmobjects.SSymbol;
 
 
@@ -30,15 +28,8 @@ public class ReceivedMessage extends ReceivedRootNode {
   }
 
   @Override
-  protected Object executeBody(final VirtualFrame frame) {
-    EventualMessage msg = (EventualMessage) SArguments.rcvr(frame);
-    boolean haltOnResolver = msg.getHaltOnResolver();
-    boolean haltOnResolution = msg.getHaltOnResolution();
-
-    if (VmSettings.TRUFFLE_DEBUGGER_ENABLED && haltOnResolver) {
-      dbg.prepareSteppingAfterNextRootNode();
-    }
-
+  protected Object executeBody(final VirtualFrame frame, final EventualMessage msg,
+      final boolean haltOnResolver, final boolean haltOnResolution) {
     try {
       Object result = onReceive.doPreEvaluated(frame, msg.args);
       resolvePromise(frame, msg.resolver, result, haltOnResolver, haltOnResolution);
@@ -65,13 +56,8 @@ public class ReceivedMessage extends ReceivedRootNode {
     }
 
     @Override
-    protected Object executeBody(final VirtualFrame frame) {
-      EventualMessage msg = (EventualMessage) SArguments.rcvr(frame);
-
-      if (VmSettings.TRUFFLE_DEBUGGER_ENABLED && msg.getHaltOnResolver()) {
-        dbg.prepareSteppingAfterNextRootNode();
-      }
-
+    protected Object executeBody(final VirtualFrame frame, final EventualMessage msg,
+        final boolean haltOnResolver, final boolean haltOnResolution) {
       Object result = onReceive.doPreEvaluated(frame, msg.args);
       future.complete(result);
       return result;
@@ -88,15 +74,8 @@ public class ReceivedMessage extends ReceivedRootNode {
     }
 
     @Override
-    protected Object executeBody(final VirtualFrame frame) {
-      EventualMessage msg = (EventualMessage) SArguments.rcvr(frame);
-      boolean haltOnResolver = msg.getHaltOnResolver();
-      boolean haltOnResolution = msg.getHaltOnResolution();
-
-      if (VmSettings.TRUFFLE_DEBUGGER_ENABLED && haltOnResolver) {
-        dbg.prepareSteppingAfterNextRootNode();
-      }
-
+    protected Object executeBody(final VirtualFrame frame, final EventualMessage msg,
+        final boolean haltOnResolver, final boolean haltOnResolution) {
       try {
         Object result = onReceive.call(msg.args);
         resolvePromise(frame, msg.resolver, result, haltOnResolver,
